+++ src/theme/variables/edgeDark.ts
@@ -342,6 +342,11 @@ export const edgeDark: Theme = {
   // Picker Color
   pickerText: palette.white,
 
+  // Native Android components like Pickers need specific text colors that
+  // contrast with the OS light or dark theme
+  nativeComponentTextLight: palette.black,
+  nativeComponentTextDark: palette.white,
+
   // Input Accessory
   inputAccessoryBackground: palette.white,
   inputAccessoryText: palette.accentBlue,